revert devops package
