--- devops/pyproject.toml
@@ -1,47 +1,0 @@
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "metta-devops"
-version = "0.1.0"
-description = "Devops support for metta projects"
-requires-python = ">=3.9"
-dependencies = [
-  "skypilot>=0.10.3",
-  "gitta",
-  "metta-common",
-  "mettagrid",
-  "metta-app-backend",
-  "wandb>=0.19.11",
-]
-
-[tool.setuptools]
-packages = [
-  "devops",
-  "devops.skypilot",
-  "devops.docker",
-  "experiments",
-  "metta.cogworks",
-  "metta.map",
-  "metta.rl",
-  "metta.sim",
-  "metta.sweep",
-  "metta.tools",
-]
-
-[tool.setuptools.package-dir]
-devops = "."
-experiments = "../experiments"
-"metta.cogworks" = "../metta/cogworks"
-"metta.map" = "../metta/map"
-"metta.rl" = "../metta/rl"
-"metta.sim" = "../metta/sim"
-"metta.sweep" = "../metta/sweep"
-"metta.tools" = "../metta/tools"
-
-[tool.uv.sources]
-gitta = { path = "../gitta", editable = true }
-metta-common = { path = "../common", editable = true }
-mettagrid = { path = "../mettagrid", editable = true }
-metta-app-backend = { path = "../app_backend", editable = true }
