+++ observatory/launch.py
@@ -6,7 +6,7 @@ import subprocess
 import sys
 from pathlib import Path
 
-from metta.common.auth.tokens import get_machine_token
+from metta.app_backend.clients.base_client import get_machine_token
 from metta.common.util.constants import DEV_STATS_SERVER_URI, PROD_STATS_SERVER_URI
 from metta.setup.utils import error, info
